Fixes for paths containing spaces, multiple .lproj locations
Changes made by Peter Ringset:

"1) The script doesn’t support paths with spaces, which causes the whole thing to crash. The for loop on line 13 will loop on incomplete file paths, as it will interpret a space as a symbol end. I found a quick solution by overwriting the IFS variable."

"2) When placing various localizable resource files in differenct places in the project directory, each .strings file is copied to every possible .lproj folder in the project directory. This results in a great deal of unused .strings files. The fix I found for this problem is to replace line 52."